[Examples] refactored both echo servers to VS2017 project format, and refactored some code
