+++ Samples/EchoServer/App.config
@@ -1,37 +1,37 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <configSections>
-    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net" requirePermission="false" />
+    <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net" requirePermission="false"/>
   </configSections>
   <startup>
-    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5"/>
   </startup>
   <runtime>
-    <gcServer enabled="true" />
+    <gcServer enabled="true"/>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
-        <assemblyIdentity name="System.Threading.Tasks.Dataflow" publicKeyToken="b03f5f7f11d50a3a" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-4.5.14.0" newVersion="4.5.14.0" />
+        <assemblyIdentity name="System.Threading.Tasks.Dataflow" publicKeyToken="b03f5f7f11d50a3a" culture="neutral"/>
+        <bindingRedirect oldVersion="0.0.0.0-4.5.14.0" newVersion="4.5.14.0"/>
       </dependentAssembly>
     </assemblyBinding>
   </runtime>
   <log4net>
     <appender name="DebugRollingFileAppender" type="log4net.Appender.RollingFileAppender">
-      <file value="activity.log" />
-      <lockingModel type="log4net.Appender.FileAppender+MinimalLock" />
-      <appendToFile value="true" />
-      <rollingStyle value="Composite" />
-      <maxSizeRollBackups value="14" />
-      <maximumFileSize value="15000KB" />
-      <datePattern value="yyyyMMdd" />
+      <file value="activity.log"/>
+      <lockingModel type="log4net.Appender.FileAppender+MinimalLock"/>
+      <appendToFile value="true"/>
+      <rollingStyle value="Composite"/>
+      <maxSizeRollBackups value="14"/>
+      <maximumFileSize value="15000KB"/>
+      <datePattern value="yyyyMMdd"/>
       <layout type="log4net.Layout.PatternLayout">
-        <conversionPattern value="%date [thread%thread] %-5level - %message%newline%newline%newline" />
+        <conversionPattern value="%date [thread%thread] %-5level - %message%newline%newline%newline"/>
       </layout>
-      <staticlogfilename value="true" />
+      <staticlogfilename value="true"/>
     </appender>
     <logger name="Main">
-      <level value="DEBUG" />
-      <appender-ref ref="DebugRollingFileAppender" />
+      <level value="DEBUG"/>
+      <appender-ref ref="DebugRollingFileAppender"/>
     </logger>
   </log4net>
 </configuration>
