+++ Samples/EchoServer/Log4NetLogger.cs
@@ -1,16 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using vtortola.WebSockets;
+using log4net;
 
 // ReSharper disable once CheckNamespace
 namespace vtortola.WebSockets
 {
     public sealed class Log4NetLogger : ILogger
     {
-        private log4net.ILog log;
+        private readonly ILog log;
 
         /// <inheritdoc />
         public bool IsDebugEnabled => this.log.IsDebugEnabled;
@@ -19,10 +15,9 @@ namespace vtortola.WebSockets
         /// <inheritdoc />
         public bool IsErrorEnabled => this.log.IsErrorEnabled;
         /// <inheritdoc />
-
         public Log4NetLogger(Type loggerType = null)
         {
-            this.log = log4net.LogManager.GetLogger(loggerType ?? typeof(WebSocketListener));
+            this.log = LogManager.GetLogger(loggerType ?? typeof(WebSocketListener));
         }
 
         public void Debug(string message, Exception error = null)
