--- Samples/EchoServer/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="log4net" version="2.0.5" targetFramework="net45" />
-</packages>
