+++ Samples/MonoEchoServer/App.config
@@ -1,0 +1,40 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+
+<configuration>
+    <configSections>
+        <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net"
+                 requirePermission="false" />
+    </configSections>
+    <startup>
+        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
+    </startup>
+    <runtime>
+        <gcServer enabled="true" />
+    </runtime>
+    <log4net>
+        <appender name="LogToConsole" type="log4net.Appender.ConsoleAppender">
+            <threshold value="INFO" />
+            <layout type="log4net.Layout.PatternLayout">
+                <conversionPattern value="%-15.15logger{1}| %message%newline" />
+            </layout>
+        </appender>
+        <appender name="LogToFile" type="log4net.Appender.RollingFileAppender">
+            <file value="activity.log" />
+            <lockingModel type="log4net.Appender.FileAppender+MinimalLock" />
+            <appendToFile value="true" />
+            <rollingStyle value="Composite" />
+            <maxSizeRollBackups value="14" />
+            <maximumFileSize value="15000KB" />
+            <datePattern value="yyyyMMdd" />
+            <layout type="log4net.Layout.PatternLayout">
+                <conversionPattern value="%date [thread%thread] %-5level - %message%newline%newline%newline" />
+            </layout>
+            <staticlogfilename value="true" />
+        </appender>
+        <root>
+            <level value="DEBUG" />
+            <appender-ref ref="LogToConsole" />
+            <appender-ref ref="LogToFile" />
+        </root>
+    </log4net>
+</configuration>
