--- Samples/MonoEchoServer/MonoEchoServer.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MonoEchoServer", "MonoEchoServer\MonoEchoServer.csproj", "{A1D77B7F-3CFB-406F-9F54-DE77E6995C5D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A1D77B7F-3CFB-406F-9F54-DE77E6995C5D}.Debug|x86.ActiveCfg = Debug|x86
-		{A1D77B7F-3CFB-406F-9F54-DE77E6995C5D}.Debug|x86.Build.0 = Debug|x86
-		{A1D77B7F-3CFB-406F-9F54-DE77E6995C5D}.Release|x86.ActiveCfg = Release|x86
-		{A1D77B7F-3CFB-406F-9F54-DE77E6995C5D}.Release|x86.Build.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = MonoEchoServer\MonoEchoServer.csproj
-	EndGlobalSection
-EndGlobal
